kino_service_grpc/requirements.txt
grpcio-tools ~= 1.62.1
sqlalchemy~=2.0.29
alembic~=1.13.1
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.29
python-dotenv~=1.0.1
alembic~=1.13.1